exampleCode/requirements.txt
beautifulsoup4
chromadb
fastembed
groq
python-dotenv
requests
streamlit
